+++ src/config.cpp
@@ -6,7 +6,7 @@ const char* default_port_range = "63000-63050";
 
 const std::uint32_t default_width = 1920;
 const std::uint32_t default_height = 1080;
-const std::uint32_t default_iterations = 1000;
+const std::uint16_t default_iterations = 1000;
 
 const float default_min_real = -1.9; // must be <= 0.0
 const float default_max_real = 1.0;  // must be >= 0.0
